Move get_srs_repr to util
