+++ src/fiat/models/base.py
@@ -18,11 +18,10 @@ from fiat.check import (
     check_vs_srs,
 )
 from fiat.gis import grid
-from fiat.gis.crs import get_srs_repr
 from fiat.io import open_csv, open_grid
 from fiat.log import spawn_logger
 from fiat.models.util import check_file_for_read
-from fiat.util import NEED_IMPLEMENTED, deter_dec
+from fiat.util import NEED_IMPLEMENTED, deter_dec, get_srs_repr
 
 logger = spawn_logger("fiat.model")
